src: move nfs4 client code into basic-client module
Motivation:
The nfsv4 client code is for testing purposes only. In addition it adds
an extra dependency which is not required by server part.

Modification:
move nfs4 client code into basic-client module.

Result:
reduced transient dependency and testing code from the core components.

Acked-by: Albert Rossi
Target: master
