+++ basic-client/src/main/java/org/dcache/nfs/v4/client/Main.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009 - 2019 Deutsches Elektronen-Synchroton,
+ * Copyright (c) 2009 - 2020 Deutsches Elektronen-Synchroton,
  * Member of the Helmholtz Association, (DESY), HAMBURG, GERMANY
  *
  * This library is free software; you can redistribute it and/or modify
@@ -54,6 +54,7 @@ import org.jline.terminal.TerminalBuilder;
 import org.dcache.nfs.v4.xdr.COMPOUND4args;
 import org.dcache.nfs.v4.xdr.COMPOUND4res;
 import org.dcache.nfs.ChimeraNFSException;
+import org.dcache.nfs.v4.CompoundBuilder;
 import org.dcache.nfs.v4.Stateids;
 import org.dcache.nfs.v4.xdr.clientid4;
 import org.dcache.nfs.v4.xdr.deviceid4;
