+++ core/src/main/java/org/dcache/nfs/v4/CompoundBuilder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009 - 2019 Deutsches Elektronen-Synchroton,
+ * Copyright (c) 2009 - 2020 Deutsches Elektronen-Synchroton,
  * Member of the Helmholtz Association, (DESY), HAMBURG, GERMANY
  *
  * This library is free software; you can redistribute it and/or modify
@@ -17,7 +17,7 @@
  * details); if not, write to the Free Software Foundation, Inc.,
  * 675 Mass Ave, Cambridge, MA 02139, USA.
  */
-package org.dcache.nfs.v4.client;
+package org.dcache.nfs.v4;
 
 import org.dcache.nfs.v4.xdr.layoutreturn_file4;
 import org.dcache.nfs.v4.xdr.nfs_fh4;
