+++ core/src/test/java/org/dcache/nfs/v4/OperationGETXATTRTest.java
@@ -2,7 +2,6 @@ package org.dcache.nfs.v4;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import org.dcache.nfs.v4.client.CompoundBuilder;
 import org.dcache.nfs.v4.xdr.COMPOUND4args;
 import org.dcache.nfs.v4.xdr.COMPOUND4res;
 import org.dcache.nfs.v4.xdr.nfs_fh4;
